requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
fastapi==0.100.1
filelock==3.15.4
h11==0.14.0
httptools==0.6.1
identify==2.6.0
idna==3.7
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.7.1
pydantic-core==2.20.1
pydantic==2.8.2
python-dotenv==1.0.1
pyyaml==6.0.1
redis==5.0.8
requests==2.32.3
segno==1.6.1
sniffio==1.3.1
spotipy==2.24.0
starlette==0.27.0
typing-extensions==4.12.2
urllib3==2.2.2
uvicorn[standard]==0.23.2
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
websockets==12.0
